requirements.txt
scikit-learn==1.1.3
numpy
flask
flask-gladiator